Split transition no verify functions in its own file
